portfolio_web/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
psycopg2-binary==2.9.11
pandas==3.0.0
numpy==1.26.4
yfinance==1.1.0
requests==2.32.5
python-dotenv==1.0.1
gunicorn==23.0.0
scikit-learn>=1.4.0
scipy>=1.10.0
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.27.2
